+++ inc/ti/spec.t.h
@@ -44,6 +44,7 @@ typedef enum
     TI_SPEC_INT_RANGE,      /* `int<:>              */
     TI_SPEC_FLOAT_RANGE,    /* `float<:>            */
     TI_SPEC_STR_RANGE,      /* `str<:>              */
+    TI_SPEC_UTF8_RANGE,     /* `utf8<:>             */
 } ti_spec_enum_t;
 
 typedef enum