+++ src/ti/spec.c
@@ -412,6 +412,7 @@ ti_spec_rval_enum ti__spec_check_nested_val(uint16_t spec, ti_val_t * val)
     case TI_SPEC_INT_RANGE:
     case TI_SPEC_FLOAT_RANGE:
     case TI_SPEC_STR_RANGE:
+    case TI_SPEC_UTF8_RANGE:
         assert(0);  /* not supported on nested definition */
         return false;
     }
@@ -505,6 +506,7 @@ _Bool ti__spec_maps_to_nested_val(uint16_t spec, ti_val_t * val)
     case TI_SPEC_INT_RANGE:
     case TI_SPEC_FLOAT_RANGE:
     case TI_SPEC_STR_RANGE:
+    case TI_SPEC_UTF8_RANGE:
         assert(0);  /* only nested so conditions are not possible */
         return false;
     }
@@ -526,6 +528,7 @@ const char * ti_spec_approx_type_str(uint16_t spec)
     case TI_SPEC_REMATCH:
     case TI_SPEC_STR_RANGE:
     case TI_SPEC_STR:           return "str";
+    case TI_SPEC_UTF8_RANGE:
     case TI_SPEC_UTF8:          return "utf8";
     case TI_SPEC_BYTES:         return "bytes";
     case TI_SPEC_INT_RANGE:
@@ -586,7 +589,8 @@ ti_spec_mod_enum ti_spec_check_mod(
             ospec == TI_SPEC_URL ||
             ospec == TI_SPEC_TEL ||
             ospec == TI_SPEC_REMATCH ||
-            ospec == TI_SPEC_STR_RANGE
+            ospec == TI_SPEC_STR_RANGE ||
+            ospec == TI_SPEC_UTF8_RANGE
         ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
     case TI_SPEC_STR:
         return (
@@ -596,9 +600,14 @@ ti_spec_mod_enum ti_spec_check_mod(
             ospec == TI_SPEC_URL ||
             ospec == TI_SPEC_TEL ||
             ospec == TI_SPEC_REMATCH ||
-            ospec == TI_SPEC_STR_RANGE
+            ospec == TI_SPEC_STR_RANGE ||
+            ospec == TI_SPEC_UTF8_RANGE
         ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
     case TI_SPEC_UTF8:
+        return (
+            ospec == TI_SPEC_UTF8 ||
+            ospec == TI_SPEC_UTF8_RANGE
+        ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
     case TI_SPEC_BYTES:
         return ospec == nspec ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
     case TI_SPEC_INT:
@@ -673,7 +682,13 @@ ti_spec_mod_enum ti_spec_check_mod(
         ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
     case TI_SPEC_STR_RANGE:
         return (
-            ospec == TI_SPEC_STR_RANGE &&
+            (ospec == TI_SPEC_STR_RANGE || ospec == TI_SPEC_UTF8_RANGE) &&
+            ocondition.srange->mi >= ncondition.srange->mi &&
+            ocondition.srange->ma <= ncondition.srange->ma
+        ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;
+    case TI_SPEC_UTF8_RANGE:
+        return (
+            ospec == TI_SPEC_UTF8_RANGE &&
             ocondition.srange->mi >= ncondition.srange->mi &&
             ocondition.srange->ma <= ncondition.srange->ma
         ) ? TI_SPEC_MOD_SUCCESS : TI_SPEC_MOD_ERR;